Warning squash

+++ onnx/defs/schema.cc
@@ -54,6 +54,7 @@ OpSchema::FormalParameterOption OpSchema::FormalParameter::GetOption() const {
 }
 
 void OpSchema::Verify(const NodeProto& node) const {
+
   // Check the number of inputs.
   if (node.input_size() < min_input_ || node.input_size() > max_input_) {
     fail_check(
@@ -89,7 +90,7 @@ void OpSchema::Verify(const NodeProto& node) const {
   }
 
   // Check the values of inputs / outputs
-  for (int in_idx = 0; in_idx < node.input_size(); ++in_idx) {
+  for (size_t in_idx = 0; in_idx < static_cast<size_t>(node.input_size()); ++in_idx) {
     if (in_idx >= inputs_.size()) {
       if (inputs_.size() > 0 && Variadic == inputs_.back().GetOption()) {
         // The last input formal parameter should be variadic.
@@ -114,7 +115,7 @@ void OpSchema::Verify(const NodeProto& node) const {
     }
   }
 
-  for (int out_idx = 0; out_idx < node.output_size(); ++out_idx) {
+  for (size_t out_idx = 0; out_idx < static_cast<size_t>(node.output_size()); ++out_idx) {
     if (out_idx >= outputs_.size()) {
         if (outputs_.size() > 0 && Variadic == outputs_.back().GetOption()) {
             // The last output formal parameter should be variadic.
@@ -454,7 +455,7 @@ void OpSchema::Finalize() {
   
   // Flag indicates whether an optional input is trailing one (there's no single or variadic
   // input behind).
-  for (int i = 0; i < (int)(inputs_.size()); ++i) {
+  for (size_t i = 0; i < inputs_.size(); ++i) {
     switch (inputs_[i].GetOption()) {
     case OpSchema::Single:
       ++max_input_;
@@ -473,7 +474,7 @@ void OpSchema::Finalize() {
   }
 
   // Calculate min/max number of outputs.
-  for (int i = 0; i < (int)(outputs_.size()); ++i) {
+  for (size_t i = 0; i < outputs_.size(); ++i) {
     switch (outputs_[i].GetOption()) {
     case OpSchema::Single:
       ++max_output_;
